Изменения:  - теперь с главной нельзя добавить новое слово, всегда надо заходить в группу (зато теперь сразу понятно что нажимать). Может быть надо добавить новую кнопку в боттомбар  - добавил в детали группы кнопку импорт  - убрал с главного экрана меню по лонг клику  - Заменил тексты и картинки на некоторых кнопках  - Переименовал Sets в Groups  - Подставил в GroupDetailsScreen PrimaryButton вместо SecondaryButton  - уберу 1.json и артифакты из коммитов
