+++ app/src/main/java/io/taptap/stupidenglish/archive/ui/MultiFloatingActionButton.kt
@@ -1,0 +1,4 @@
+package io.taptap.stupidenglish.archive.ui
+
+class MultiFloatingActionButton {
+}
