+++ app/src/main/res/values/strings.xml
@@ -30,20 +30,20 @@
     <string name="word_motivation_message">Let\'s make some sentences with your words</string>
     <string name="word_motivation_confirm">Let\'s do it!</string>
     <string name="word_motivation_decline">Nope</string>
-    <string name="word_fab_text">Add word</string>
-    <string name="word_group_title">Sets</string>
+    <string name="word_fab_text">Open group</string>
+    <string name="word_group_title">Groups</string>
     <string name="word_group_button">View all</string>
-    <string name="word_group_add_group_title">New set</string>
+    <string name="word_group_add_group_title">New group</string>
     <string name="word_empty_list_title">List is empty</string>
-    <string name="word_empty_list_description">Add some words to the set</string>
-    <string name="word_group_menu_title">Set\'s menu</string>
+    <string name="word_empty_list_description">Add some words to the group</string>
+    <string name="word_group_menu_title">Group\'s menu</string>
     <string name="word_remove_group_button">Remove</string>
     <string name="word_delete_message">Are you sure?</string>
     <string name="word_recover">Recover</string>
     <string name="word_minifab_manual">Manual</string>
     <string name="word_minifab_import">Import</string>
     <string name="word_get_user_error">Error while getting user</string>
-    <string name="word_menu_open">Open set</string>
+    <string name="word_menu_open">Open Group</string>
     <string name="word_menu_add_word">Add word</string>
     <string name="word_menu_flashcards">Flashcards</string>
     <string name="word_menu_learn">Learn</string>
@@ -70,6 +70,7 @@
     <string name="grdt_delete_message">Are you sure?</string>
     <string name="grdt_recover">Recover</string>
     <string name="grdt_add_word">Add word</string>
+    <string name="grdt_import_words">Import</string>
     <string name="grdt_flashcards">Flashcards</string>
     <string name="grdt_learn">Learn</string>
     <string name="grdt_share">Share</string>
