+++ core/uikit/src/main/java/io/taptap/uikit/group/GroupItemRow.kt
@@ -1,6 +1,5 @@
 package io.taptap.uikit.group
 
-import android.util.Log
 import androidx.compose.animation.core.animateFloatAsState
 import androidx.compose.foundation.ExperimentalFoundationApi
 import androidx.compose.foundation.Image
@@ -24,13 +23,10 @@ import androidx.compose.material.Text
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Alignment.Companion.CenterVertically
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
 import androidx.compose.ui.draw.scale
 import androidx.compose.ui.draw.shadow
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.ColorFilter
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
@@ -61,7 +57,6 @@ fun GroupItemRow(
     onMove: (ItemPosition, ItemPosition) -> Unit,
     onPlusClicked: () -> Unit = {},
 ) {
-    Log.d("LOGLOGLOG", "way = GroupItemRow")
     Column(
         modifier = Modifier
             .fillMaxWidth()
@@ -168,6 +163,7 @@ private fun PlusGroupItem(
             color = MaterialTheme.colorScheme.primary,
             modifier = Modifier
                 .size(44.dp)
+                .shadow(1.dp, RoundedCornerShape(44.dp))
                 .clickable {
                     onPlusClicked()
                 }
@@ -206,6 +202,7 @@ private fun GroupItem(
             fontSize = 28.sp,
             modifier = Modifier
                 .size(56.dp)
+                .shadow(1.dp, shape = RoundedCornerShape(56.dp))
                 .combinedClickable(
                     onClick = { onGroupClicked(group) },
 //                    onLongClick = { onGroupLongClicked(group) },
@@ -234,24 +231,24 @@ private fun GroupItem(
     }
 }
 
-//@Preview(showBackground = true)
-//@Composable
-//fun GroupItemRow() {
-//    StupidEnglishTheme {
-//        GroupItemRow(
-//            title = "Groups",
-//            button = "Add",
-//            currentGroup = null,
-//            onButtonClicked = {},
-//            list = emptyList(),
-//            onGroupClicked = {},
-//            onGroupLongClicked = {},
-//            onMove = { a, b ->
-//
-//            }
-//        )
-//    }
-//}
+@Preview(showBackground = true)
+@Composable
+fun GroupItemRow() {
+    StupidEnglishTheme {
+        GroupItemRow(
+            title = "Groups",
+            button = "Add",
+            currentGroup = null,
+            onButtonClicked = {},
+            list = emptyList(),
+            onGroupClicked = {},
+            onGroupLongClicked = {},
+            onMove = { a, b ->
+
+            }
+        )
+    }
+}
 
 @Preview(showBackground = true)
 @Composable
